+++ tailwind.config.js
@@ -87,6 +87,8 @@ module.exports = {
         },
         scale: {
             '-1': '-1',
+            75: '.75',
+            100: '1',
         },
         flex: {
             1: '1',
